Restructure project with src and tests folders
