+++ index.html
@@ -13,7 +13,7 @@
     </div>
   </nav>
   <div id="portfolio" class="card"></div>
-  <script src="portfolio.js"></script>
+  <script src="src/portfolio.js"></script>
   <script>
     document.addEventListener('DOMContentLoaded', function() {
       const p = typeof portfolio !== 'undefined' ? portfolio : {};
